Add support for count argument in `evil-repeat'
An explicit count for evil-repeat replaces the count of the command to
be repeated. We do this with a special pre-command-hook which works
for simple commands but an operator has to be handled differently. In
that case the interactive handler for operators has to use
`evil-repeat-count' (if non-nil) as count for the following motion.