Make motion selectors more Emacsish, basic implementation for sentences and paragraphs
Selector functions now work in the spirit of `forward-word' and
friends. If given a positive count they move point to the first
character after the count-th word, if given a positive count they move
point to the first character of the count-th previous word. The return
value is always the number of objects that could not be moved.

Unfortunately, most Emacs functions do not follow these interface
strictly, especially the return value. The implementations for paragraphs and sentences do currently not follow ViM specifications.