Make "n"/"N" open invisible areas
Add a feature for the search next/previous commands "n" and "N" so
that if `search-invisible' is `open', these commands are able to open
invisible areas when the match enters such an area. They also
automatically close a previously opened area (provided that the
subsequent "n/N" is pressed within `evil-flash-delay' and no
"non-search" commands are executed in the meantime).