Advise `display-buffer'
When a buffer is created in a low-level way, it is invisible to Evil
(as well as other globalized minor modes) because no hooks are run.
This is appropriate since many buffers are used for throwaway
purposes. Passing the buffer to `display-buffer' indicates otherwise,
though, so add an advice that initializes Evil in such buffers.