Add Ex grammar
Add the variable `evil-ex-grammar' for defining the minilanguage
of Ex, which is passed to `evil-parser'. The function `evil-ex-update'
refreshes the evaluation tree; when applicable, it also refreshes
parse-specific variables like the current command, argument, etc.

Furthermore, it is now possible to define key sequences that execute
a command immediately when entered. Such shortcuts go in `evil-ex-map'
(the completion bindings have been moved to `evil-ex-completion-map').

Note that some of the variable names have been simplified: e.g.,
from `evil-ex-current-cmd' to `evil-ex-command' and from
`evil-ex-current-cmd-force' to `evil-ex-force'. `evil-ex-message'
has been renamed to `evil-ex-echo' for consistency with `evil-echo'.