Make `evil-line' move over folded lines (fixes #85)
Formerly `evil-line' moves point down several visible lines. But if
the target line is a folded one, the buffer line that corresponds to
this position is usually the first of the folded lines. Thus an
operator would only be applied up to the first of the folded lines.

The new behavior is to move one visible line further and, if
successful, move back to the end of the previous line. If the target
line is unfolded the result is the same as before. But if the target
line is folded this effectively moves point to the end of the last of
the folded lines, hence including all folded lines in the
corresponding operator range.