Make `evil-insert-newline-below' open line below folded lines
If `evil-open-below' is executed on a folded line, the insertion of
the new line took place below the first of the folded lines. The
correct behavior is to open the new line below all folded lines. This
is achieved by moving to the first line below the folded lines with
`next-line' and then moving back to the end of the last folded line.