Use `move-beginning-of-line' and `move-end-of-line'
Replace `beginning-of-line' with `move-beginning-of-line' and
`end-of-line' with `move-end-of-line'. These functions recognize
hidden regions such as folds, moving point to the end of the line
as it is displayed. Thus, the fold can be properly selected by the
motion's type, instead of by the motion itself or the operator
applied to it. This fixes #88.