Do not load "goto-chg.el" from "evil-integration.el"
If the goto-chg.el package is available and installed properly it is
autoloaded (or should be loaded manually). The corresponding motions
g; and g, are now *always* bound to `goto-last-change' and
`goto-last-change-reverse', resp., and work if goto-chg.el is
available, otherwise an error is raised. The old implementation loaded
"goto-chg.el" automatically from "evil-integration.el". But even if
this succeeded the corresponding bindings for g; and g, would not have
been installed (unless "goto-chg.el" had been loaded manually before
evil), because "evil-maps.el" is loaded *before*
"evil-integration.el", thus the loading of "goto-chg.el" in the latter
file is useless.

The package "lib/goto-chg.el" coming with evil is loaded automatically
for "make emacs" and "make terminal".