`evil-esc' modifies keyboard macro being defined correctly
If a keyboard macro is being recorded the ESC event is put into the
macro literally which leads to a wrong behavior when the keyboard
macro is replayed and the ESC key must be mapped to 'escape. This
patch temporarily stops the recording of the macro, inserts an
'escape event and the resumes the recording afterwards.