Add transformation function for vim-style regexp
The new function `evil-transform-regexp' converts Vim-style backslash
codes to the corresponding character classes, e.g., \d is transformed
to [[:digit:]]. The following codes are currently supported: d, D, s,
S, x, X, o, O, a, A, l, L, u, U. The codes w and W are the same as in
Emacs so they are supported automatically. Furthermore the Emacs code
\s is overwritten to be interpreted as [[:space:]]. As a compromise,
the new code \y is transformed to the Emacs \s and can be used in
Vim-style regular expression in place of the Emacs \s.