Move point on `evil-ex-search-next' forward at eol
When point is at eol and `evil-move-cursor-back' is t then point is
(on non-empty lines) moved back one character. If a search command /
matches the end of a line point would be moved back and a subsequent
execution of `evil-ex-search-next' (command n) would therefore find
the same end of line. This patch changes `evil-ex-search-next' to test
for the special case of being at eol and ensures point moves forward.
This addresses #140.