Make `evil-delete-line' yank deleted block correctly
When `evil-delete-line' is called from block visual state then all
lines of the block should be deleted to the end of the line. The old
implementation called `evil-apply-on-block' to delete the end of each
line. Unfortunately this appended the deleted text of each line as a
single item to the kill ring instead of appending the deleted text as
one block. This patch makes `evil-delete-line' in block mode
equivalent to the sequence $d, which extends the block selection to
the end of each line and then deletes the whole block (in fact, in
normal-state `evil-delete-line' is a synonym for d$). This places the
deleted text correctly as a block to the kill ring.