Make key-bindings in `evil-read-key' configurable
The former implementation hard coded several keys with special meaning
in `evil-read-key'. The new implementation introduces a new keymap
`evil-read-key-map' that is active during `evil-read-key'. All
hard-coded keys are now implemented as bindings in this keymap.

Furthermore in the case of a missing key-binding or a non
self-insert-command (such as mouse events) an error is raised.

This fixes #149.