Make `evil-write-all' corresponding to ViM :wall
The command :wall should never asked for a buffer being saved. Instead
it should save all buffers that are visiting a file and that are not
read-only. With a bang ! also read-only buffers are saved.

The old behavior of evil was to ask for each buffer if it should be
written to its file if :wa has been called without bang.

The new behavior now mimics Vim. The call with bang :wa! has not been
changed (it has already been correct).