Register visual hooks only in buffers with activated evil
The function `evil-visual-activate-hook' is registered must be
registered as `activate-mark-hook'. This should happen only localy in
buffers where evil is activated. Otherwise the activation of the
region could enable evil (in visual state) even in buffers that should
not have evil enabled (e.g., the minibuffer).

The old behavior registered this hook globally. Not it is a
buffer-local hook. This fixes #150.