Contract range in `evil-operator-range' before setting its type
Before setting the type of the result range (if this should happen)
the range must be contract and being expanded after setting the range.
Without the contraction the subsequent expansion would not do anything
if the original range has already been expanded (which is the case,
e.g., for visual ranges). This fixes #163.