Add `evil-exit-visual-and-repeat'
This is a special command to be bound in visual state. It should be
used if some command can be called in visual state but should be
actually called in normal state. It simply exits visual state and is
ignored by the repeat system. After stopping visual state it resends
the event by which it has been invoked. Thus the same event that
invoked `evil-exit-visual-and-repeat' is resend but this time in
normal state. The command that should be executed must be bound to the
same event but in normal state.