Support the last insertion register ". .
`evil-track-last-insertion' as a post command hook calculates the
last piece of insertion.  If two successive commands insert
continuous text, the text is treated as a piece of insertion like
successive `self-insert-command's fix up their undo information.
It uses `evil-last-undo-entry' and `evil-last-insertion-command' to
determine whether we have some successive insertion commands.

The value of the register is not updated until the current
insert-state session ends.  `evil-current-insertions' has
information about insertions in the current insert-state session,
which is saved to the register value at the end of the session.

It is a read only register (`evil-set-register' does not change its
value).

This addresses issue #196.