Fix appending of lines to a register
If some text is appended to a register using "aY or similar and if
either to previous content or the new content has a non-trivial
yank-handler, the text is appended as new lines to the register. The
yank handler for the *whole* register content is changed to linewise,
possibly adding newline characters, independently of the previous
type.

This fixes #197.