Remove adjustment of point after the final newline of a buffer
This patch makes an empty line appear after the final newline in a
buffer. This is different to Vim which shows no empty line beyond that
final newline, but this behavior is more consistent with Emacs. Some
Emacs modes depend on that final newline not being hidden, so the old
behavior causes some trouble when evil is used in such modes.

Besides removing the line adjustment, several commands (change,
delete) had been modified. These commands cannot rely on each line
ending in a newline anymore, but must deal with the situation that
some line does not end in a newline (in particular, the last visible
line of buffer) explicitly. Note that most other commands do not
interact with empty lines at all, so they do not need to take care of
that empty line at the end of the buffer.

This addresses #190.
This addresses #198.
This addresses #209.
This addresses #221.