Set the region instead of visual selection for ex commands
We now set the region explicitly to match the ex-range when an ex
command is called. We do not activate visual state anymore.
Furthermore we set the region even if a visual selection is currently
active. This is because ex commands should always ignore the visual
selection, unless it is specified explicitly using register ranges
'<,'>.