Make `evil-replace` handle block selections with TABS correctly.
The old behavior, using `evil-apply-on-block`, considered a TAB as a
single character, thus replacing the TAB with exactly one character.
This leads to an unexpected visual impression, because one TAB usually
represents several spaces.

The new implementation uses `evil-apply-on-rectangle` and deals
explicitly with the columns covered by the block in each line. TABS
are converted to spaces if required.