Use markers for identifying lines in `evil-ex-substitute` (fix #297).
The old implementation used line numbers to restrict the substitution
to a certain range of lines. This may not work if the replacement text
adds line breaks. The new implementation sets one marker at the
beginning of each line in the specified range before the substitution
is done. Then the substitution is performed in each line that has such
a marker. Since markers are moved on insertion of text, this will
correctly handle inserted line-breaks.