optionally rebalance windows after creation/deletion (fix #358)
If the user option `evil-auto-balance-windows` is non-nil then Evil's
functions that create new windows or delete existing windows
automatically rebalance the children of the new/deleted window's
parent window. The default value is t to resemble Vim's default
behaviour. Note that Emacs' built in functions do not cause
rebalancing of the windows.