make backward search more Vim compatible
If point is on a match but not on its first character, then backward
search in Vim finds this match. However, standard Emacs search
functions (in particular `re-search-backward`) only match if the end
of the match is *before* the starting point. This change modifies
Evil's search function so that it behaves like vim using an additional
search step.