Do not copy original keymap when making overriding map
When calling `evil-make-overriding-map', do not specify 'copy because doing so
would cause KEYMAP been "frozen" in whatever state it is, i.e., any binding on
the original map after this point will not take effect. For example, if user
(or some other package) binds some key in `dired-mode-map' after Evil has been
loaded by using `(define-key dired-mode-map "SOME-KEY" 'SOME-COMMAND)' that
binding will never take effect because it is the copy of `dired-mode-map' rather
than the original map that is in use.