Move `evil-mode' definition to evil.el.
Do this so that when Emacs generates an autoload for `evil-mode', it
references the correct file to later load all of evil.

We do that so that we don't have to manually override Emacs'
generation of the ;;;###autoload cookie.

We do THAT so that we can let Emacs generate the appropriate autoloads
to enable the globalized minor mode via `custom-set-variables'.