Move point to the partial match in an unsuccessful search
In an incremental search, move point even if the search matches only part of the
searched string. This is the behavior of the Emacs `isearch', and is more
convenient because incremental search has been frequently used to quickly move
the point, in which case it's too easy to have a typo. It would be really
annoying if we have to restart the search only because of a minor typo.