maps: Make insert-state bindings "optional"
The variable evil-insert-state-bindings holds all insert-state bindings except ones for <delete>, <escape>, and evil-toggle-key.

The variable evil-disable-insert-state-bindings prevents the bindings from being bound when evil is loaded.

New functions evil-add-insert-state-bindings and evil-remove-insert-state-bindings add or remove these bindings as long as they don't conflict with other changes made by the user.

The motivation is that some users might prefer Emacs bindings in insert state over the vim ones, and this makes it easy to accomplish that goal.