Add `evil-text-object-change-visual-type` customization option (fix #607)
Setting this variable to nil causes text-objects not to change the
type of the current visual selection.

If the resulting visual selection in linewise and the first selected
character is a newline character, that character is excluded from the
selected range (otherwise the linewise selection would include the
corresponding line, which is usually not desired). This is similar to
the common exclusive-linewise rule.