Add evil-define-minor-mode-key and supporting funcs
evil-define-minor-mode-key is similar to evil-define-key, except
it associates bindings with a minor-mode symbol instead of a
specific keymap. This has the advantage of automatically picking
up on when a minor-mode is enabled or disabled through standard
Emacs mechanisms, instead of requiring a call to
evil-normalize-keymaps to activate the binding after the mode is
toggled.

This is meant to supplement usage of evil-define-key.