Fix initial space in command-window
The initial content of the last command window line was a single space. The
reason was that otherwise the line prefix is not drawn.

This is problematic because it causes a single space to be appended to entered
commands. While ex commands usually handle this correctly, it is important
for search.

The new implementation ensures that the command window ends in an empty line,
the input is expected on the second to last line.