Replace "P" interactive code by "<c>"
The interactive code "P" returns a *raw* prefix argument, which might
be a number but might also be something else (e.g. if only C-u is
used). In contrast, "<c>" always returns a number (or nil if no prefix
argument is given). Because most function depend on the count argument
to be either `nil` or a number, using "<c>" instead of "P" is
appropriate.