evil-core.el: Factor function out of evil-define-key
and into a new function evil-define-key*. evil-define-key* tries
bind the keys on invocation which leads to more predictable behavior.
For example, if there is a typo in the keymap then using
evil-define-key will never bind its keys, whereas evil-define-key*
will signal an error right away. It is also a function instead of
a macro which is much more predictable inside of loops and functions
when the file is compiled and the macros are expanded.