Tweak undo behavior to fix #594
In v25.1, emacs changes the way undo boundaries are automatically
removed to combine commands into single undo steps. #594 is the
result of more aggresive removals of undo boundaries.

To fix this, make evil more active in inserting undo boundaries
when entering and exiting insert and replace state by creating
a boundary automatically if none exists.

A slightly different issue in #594 is that the fine option of
evil-want-fine-undo was not always working correctly. The reason
is that it removed the first undo boundary after exiting insert
state regardless of whether it made sense to do so
(i.e., regardless of whether there was a change command). This
commit removes the fine option and simplifies the meaning of the
variable to mean aggregate commands like vim or use Emacs
heuristics.