evil-core.el: Add <leader> functionality
Add evil-send-leader and evil-send-localleader which will push the symbol
<leader> or <localleader> into the unread-command-events list. This allows one
to bind a key to evil-send-leader to make it a leader key and activate bindings
like the following.

(defun hi ()
  (interactive)
  (message "hi"))
(define-key evil-normal-state-map (kbd "<leader> h") 'hi)

For convenience, evil-set-leader can be used as follows

(evil-set-leader 'normal (kbd "C-c"))

An optional argument makes it a localleader.