prevent warning when accessing search history
Making a variable buffer-local within the BODY of a `let' generates a warning if that
variable is currently let-bound (i.e. it appeared in the `let' VARLIST). Such warnings are
only issued once per symbol, in order to avoid harassing the user too much.

Within evil-search-incrementally, 'isearch-search-fun-function is let-bound, and
within the body of that let, isearch makes that same variable buffer-local (see
minibuffer-history-isearch-setup). As a result, upon starting Emacs in evil-mode, the
first time you attempt a search (/), accessing the search history (M-p) causes the
following warning to display in the echo area:

  Making isearch-search-fun-function local to *Minibuf-1* while let-bound!

Subsequent searches are unaffected, since these warnings are only displayed once per
symbol. If the first search doesn't try to access search history, it too is unaffected.

This was reproducible with a minimal init.el

    (package-initialize)
    (require 'evil)
    (add-hook 'after-init-hook #'evil-mode)

and the following steps:
 - start Emacs
 - type /
 - type M-p

In order to prevent such warnings, make 'isearch-search-fun-function buffer-local before
entering the `let'.