Update evil-scroll-line-up and ..-down
Adding an "&optional" keyword indicates that the count argument
isn't required.

The docstrings were updated to match the changes. With an addition
of stating that a negative count scrolls in the opposite direction.

A 0 count argument caused the evil-scroll-line-up / ..-down
commands to stop scrolling, this resets a 0 count argument to 1.