Fix regression in evil-ex-substitute with nxml-mode
nxml-mode somehow changes the match data during the course of using
evil-ex-substitute with the gc flags. Make evil-ex-substitute very careful about
using the right match-data for moving point and performing replacements. This
should not be necessary, but I guess we can't rely on modes to be careful about
not overwriting the match data randomly.

Fixes #859