Try to handle special cases better in evil-ex-substitute
One special case occurs with the ex commands ":s/^/ #/" or ":'<,'>s/^/ #/". The
range passed to evil-ex-substitute includes the final newline which means that
the zero-length regex "^" will match on the line below the intended range. This
is now checked for explicitly.

Another occurs using the regex "$", as in ":%s/$/ TEST". In this case, after the
replacement occurs point is still looking at the end of the same line. To test
for this case, when there is a zero-length match we test the next search to see
if point won't move following the next (successful) search. If so, we move point
explicitly.

Finally, this fixes an issue where we could inadvertently move point past
end-marker. This wasn't a big deal, but it did print a message in the *Messages*
buffer.

Fixes #881