Remove redundant function level requires
Now that `shell' has been required at the top level, we don't need to require it
inside any functions in the same file. To calm down byte compiler, `require'
statements have to be placed outside of any function because they are not parsed
at compile time if they are inside functions.