Use initial states of parent major modes
Teach evil-initial-state to look at aliases for a mode when they exist and to
handle nil for modes

Search parent modes (and their aliases) for defined initial states in
evil-initial-state-for-buffer.

One effect is that

(evil-set-initial-state 'special-mode 'motion)

now makes motion state the default for all major modes that derive from special
mode and don't have defaults set for them.