Add evil-define-key-with-key-func
This is an anaphoric version of evil-define-key which allows one to transform
all of the key descriptions according to a "key-func". A useful example is

(evil-define-key-with-key-func (kbd key) STATE KEYMAP BINDINGS)

which expands to a standard evil-define-key call after applying kbd to every key
in BINDINGS. More info is in the docstring.