Don't move point following ex-substitute search
The current behavior isn't consistent with vim (at least in my tests). It's also
annoying to have point jump to weird locations following an aborted search.

For example, with
:%s/a/b/c

If you abort the search with C-g before making replacements, point jumps to the
beginning of the region which is the beginning of the buffer in this case. This
is odd, because nothing happened but you lose your place in the buffer.