Fix scrolling up/down in the presence of a header line
When `point` is at the first line of a window and a header line is present,
`posn-point` will return `nil` since `point` cannot be within a header line. To
work around this issue, when a header line is detected at the position of
`point` this means that `point` is at the first line visible in the window so
we can use `move-to-window-line` in this case.