Better integration of window manipulation commands
Commands like `evil-window-rotate-upwards` and `evil-window-move-very-top` only
consider the buffer in each window when manipulating the window tree, but if a
window has specific window properties that need to be maintained (like those in
`window-persistent-parameters`) that information is lost.

Refactor the window manipulation commands to use `window-state-get` and
`window-state-put` to retrieved and restore the state of the window tree so
that window properties in each window are maintained.

Also, integrate the window manipulation commands with side windows, i.e. windows
with a non-nil `window-side` window parameter, by only manipulating the part of
the window tree that are not side windows.

* Add the macro `evil-save-side-windows` which removes all side windows on
  entering the macro environment and restores the side windows leaving.

* Add the function `evil-move-window` which does the work of
  `evil-window-move-far-left`, `evil-window-move-very-top`,
  `evil-window-move-far-right`, and `evil-window-move-very-bottom`. These
  functions were all identical, the only difference being which window was used
  to restore the `selected-window`.

* Use `window-state-(get|put)` to get the window tree and to restore it. This
  makes `evil-get-buffer-tree` and `evil-restore-window-tree` unnecessary and
  they are removed. Note that `window-state-(get|put)` also handles window
  parameters.

* In addition to ignoring side windows when moving a window, ignore them in
  `evil-window-rotate-downwards` and `evil-window-rotate-upwards`.