Fix bound of wrapped search in evil-search-function
evil-search-function performs an initial search starting from point.  If
that search fails and search wrapping is enabled, evil-search-function then
performs a second search, this time from the start of the buffer.  However,
evil-search-function was using the same bound for both searches.  This
commit changes evil-search-function to use the first search's starting
position as the search bound for the second search.

This commit fixes issue #843.

* evil-search.el (evil-search-function): Use the initial search's start as
the bound for the second search when searching wraps.