evil-show-registers: Show macros stored as vectors
When displaying register contents, evil-show-registers substitutes the
empty string for non-string values.  If a register contains a macro, then
the value is a key sequence, and a key sequence may be represented by
either a string or a vector.  The latter would be substituted by the empty
string.

This commit checks for vector values and uses key-description to display
them.

* evil-commands.el (evil-show-registers): If a register contains a vector,
display it by calling key-description.