Support recursive minibuffers in Ex
When `enable-recursive-minibuffers` is non-nil it is
possible to enter another minibuffer at a lower editing
depth from within the Ex minibuffer.

Since `evil-ex-teardown` is added to the global value of
`minibuffer-exit-hook`, it will only be called in the
minibuffer at the lower level when it exits instead of in
the Ex minibuffer since it is removed from
`minibuffer-exit-hook` when it is called.

But this means that `evil-ex-update` will not be properly
cleaned up since it is added to the local value of
`after-change-functions` in the Ex minibuffer.

After completely exiting all minibuffers and then entering a
minibuffer that is not an Ex minibuffer, `evil-ex-update`
will still be in `after-change-functions` due to how
minibuffers are
named (https://www.gnu.org/software/emacs/manual/html_node/elisp/Intro-to-Minibuffers.html)
and assuming that the initial Ex was entered when the
recursive editing depth was 0.

Instead of adding `evil-ex-teardown` to the global value of
`minibuffer-exit-hook`, this commit adds it to the local
value of `minibuffer-exit-hook` in the Ex minibuffer so that
`evil-ex-update` is properly removed regardless of how many
recursive minibuffers there are.