Delete window on q if created to hold evil-list-view-mode buffer
Windows created specifically to hold evil-list-view-mode buffers seem a lot like
other temporary windows in Emacs, e.g. the ones created by describe-function.

These tend to define q to mean quit-window, rather than kill-this-buffer; The
quit-window command has the nice property that if a window was created to hold
the buffer in question, that window will be deleted – but if an existing window
was repurposed to hold the buffer, the effect is the same as bury-buffer.

Note that this is also in line with the default behaviour of Vim, where pressing
q in the marks list restores the previous window layout.