Fix behavior of evil-scroll-up and evil-scroll-down again
The PR #1154 is supposed to fix the issue about evil-scroll-up
and evil-scroll-down. It works nicely in most scenarios.
However, when the header line is really tall,
evil-header-line-height will give an incorrect result,
making the bug that is supposed to be fixed by the mentioned PR
reappear again.

This PR switches evil-header-line-height to Emacs' built-in
window-header-line-height which corrects the mistake.