Implement Vim Ctrl-f in minibuffer.
Open apropriate command windows for Ctrl-f while in minibuffer for an ex command
or a search. As all ways to abort a minibuffer do not return and restore the
active buffer, we have to set the content of the minibuffer and exit normally to
trigger the apropriate action. Also restoring the window configuration is
necessary so minibuffer-selected-window is not changed.