Fix issue 1295: failed visual-paste from registers >1
The `kill-ring` let binding in `evil-visual-paste` meant that
`evil-get-register`'s view of `kill-ring` (used in `evil-paste-before`
and `evil-paste-after`) was limited to only entry `0` (and whatever
`evil-delete` may have added to rebound `kill-ring`). Therefore,
registers >1 were not actually visible to functions that needed
them. Condensing the `let*` binding so `kill-ring-yank-pointer` is
bound directly to `(list (current-kill 0))` was chosen as a simple and
effective method to fix the bug.

Test added to `evil-test-visual-paste` to ensure higher number
registers are available to visual-paste.