Default `exwm-mode` to Emacs state
EXWM buffers display X windows, as such it is not useful to be in an Evil state as there is no text to manipulate. Being in normal state causes some problems, for example the simulation key feature does not work well in normal mode (as documented at https://github.com/ch11ng/exwm/wiki/Notes-to-Spacemacs-users#simulation-keys).