Consider line-move-visual and last-command in evil-line-move
line-move is hard-coded such that it will not perform visual movements
when last-command is next-line or previous-line. If
line-move-visual is set, then we are trying to move the point visually,
but this will not happen if last-command is next-line or
previous-line. Thus, we set last-command to nil in these cases.