Do not try to switch file/buffer when jumping inside same file
The vast majority of jumps will end up within the current opened
file/buffer, simply because usually users do more movement inside a
buffer rather than across buffers.

There is even a special case when evil-jumps-cross-buffers is set to
nil, which causes jumps never ever switch buffers.

So, let's do a small optimization, and not try to call find-file or
switch-to-buffer if we know it is unnecessary.

As a side-effect, this should also help to users to stumble less upon
https://github.com/emacs-evil/evil/issues/1504 bug (it has its own
upstream fix, but until that landed, this small optimization is known to
help as well).