Add option to stop updating X PRIMARY selection with the current visual region
Vim user can select a region in visual mode and press the key `p` to
replace the selected text.

But in GUI Emacs, the selected text might not be replaced when:
- `select-enable-primary` is `t`
- Third plugins override `interprogram-paste-function` (xclip.el, for example)
- Clipboard managers syncronise data between PRIMARY and CLIPBOARD selection

It's because the visual commands automatically insert the content of visually
selected region into X PRIMARY selection. But paste commands call the api
`current-kill` which calls `interprogram-paste-function`.

Value of `interprogram-paste-function` is `gui-selection-value` which
might return CLIPBOARD or PRIMARY selection.