Restore `evil-select-an-object` API
Before commit

    b799fca7 Add `evil-select-an-unrestricted-object`

evil-select-an-object was used for unrestricted selection. b799fca7
made this function restrict selection to a line, and introduced a new
function evil-select-an-unrestricted-object that works same way as
evil-select-an-object before.

But evil-select-an-object is an API that might be used by people in
custom configuration, and changing behavior of the function would result
in silent breakage.

It is also worth noting that documentation at doc/source/extension.rst
and doc/build/texinfo/evil.texi still mentions the older unrestricted
behavior.

So let's restore the original behavior of evil-select-an-object by:

1. Renaming evil-select-an-object → evil-select-a-restricted-object
2. Renaming evil-select-an-unrestricted-object → evil-select-an-object