Make sure visual markers are up to date in evil-visual-range
With composed characters in the buffer, such as with prettify-symbol-mode
active, emacs may adjust point after the post-command-hook is run. In this case,
evils visual markers do not get updated prior to say evil-delete and the deleted
region does not agree with what is seen visually (only part of the composed
characters are deleted). Forcing a last update using evil-visual-refresh
in evil-visual-range resolves this issue.