Account for composed characters in inclusive type
Previously the next position in an evil-expand operation was assumed to be one
position to the right (or left), but this position might be invalid if there are
composed characters in the buffer. This can make it difficult to delete a
composed character with vd for example, because the visual range is not expanded
properly to account for the composition.

This commit adds two new functions, evil-next-adjusted-pos and
evil-prev-adjusted-pos, which are aware of composed characters. The new
functions are then used in the inclusive type in evil-types.