Make evil-range aware of composed characters
A composed character is a way of visually replacing a string of characters with
a single glyph. If beg in the call to evil-range lies in this hidden sequence,
adjust the position to the beginning of the sequence. Similarly, make sure end
comes at the end of any such hidden sequence. The purpose is to prevent evil
from unintentionally modifying part of a hidden sequence of characters.