Fix up the cmds that scroll relative to cursor
Reverts commit bbe92d1 (Handle scroll margin in scroll-to-top/bottom
functions (re #618), 2016-03-31) which introduced a bug in
evil-scroll-bottom-line-to-top (due to recenter argument becoming
negative.) recenter itself has clamped to scroll_margin for at least 18
years now, so it is unclear why that commit was even necessary.

Also fixes the docstring for evil-scroll-top-line-to-bottom and avoids
scanning down from (point-min) to the line the point is already at.