Simplify temporary-goal-column logic
In `evil-ensure-column` there is no point in normalizing
`temporary-goal-column` a second time after BODY, since that variable is
used to store the saved column that need not be overwritten. Also
tracking EOL after `evil-end-of-line` is already handled as it fakes
being a `next-line` command.

Adds a test for commit 7aa71576d258b9546238f6dbd4cf5ca0a1b70088.