Make visual block selection after j/k match Vim
Vim only extends the visual block selection to the temporary goal column
if tracking EOL. Consider

    [a]bc
    a

When entering Visual block mode with "C-v", going to the "c" in "abc"
and then to the next line with "fcj", the selection should only include
the two "a":s, and not "bc". Contrast this with when "$" is executed.