Remove some unnecessary evil-normalize-position:s
All Emacs functions that take positions as arguments such as goto-char
already do the normalization performed by evil-normalize-position. The
only case where it would be necessary is when a marker that may be
mutated is stored. Therefore it would probably be a good idea to replace
evil-normalize-position with conditional marker-position altogether.