Wrap just show-paren-data-function in advice
Instead of positioning the point and narrowing correctly for the
entirety of show-paren-function, just do it for the
show-paren-data-function call. This avoids duplicating the logic for
removing stale overlays, which had to support both the new and old names
of the once renamed show-paren--overlay variables.

Not narrowing during show-paren-function also fixes an issue with the
"mixed" show-paren-style, which uses pos-visible-in-window-p, see #1739.

Closes #1739